requirements.txt
Flask==2.3.0
Flask-CORS==4.0.0
mysql-connector-python==8.1.0
openai==0.27.8
requests==2.31.0
python-dotenv==1.0.0